# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Stronghold API
  description: Feed monitoring and notification system API
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Local development server
paths:
...
  /api/feed-filters/{id}:
    get:
      summary: Get a feed filter by ID
      operationId: getFeedFilter
      tags:
        - Feed Filters
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Feed filter details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedFilterResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Update a feed filter
      operationId: updateFeedFilter
      tags:
        - Feed Filters
      parameters:
        - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeedFilterRequest'
      responses:
        '200':
          description: Feed filter updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedFilterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Delete a feed filter
      operationId: deleteFeedFilter
      tags:
        - Feed Filters
      parameters:
        - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Feed filter deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

  # Feed Author Filters
...
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: Resource ID
      schema:
        type: integer
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # Filter Keys
...
    FeedFilterRequest:
      type: object
      properties:
        name:
          type: string
        feed_id:
          type: integer
        category_id:
          type: integer
        notifier_id:
          type: integer
      required:
        - name
        - feed_id
        - category_id
        - notifier_id
    FeedFilterResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        feed_id:
          type: integer
        feed_name:
          type: string
        category_id:
          type: integer
        category_name:
          type: string
        notifier_id:
          type: integer
        notifier_name:
          type: string

    # Feed Author Filters
...